DNN-4893 Fix broken jQuery UI dialog style
When using the jQuery UI dialog, the "background" overlay appears over the dialog. This happens because the default.css gives the overlay a z-index, but does not give the dialog a z-index (instead, it gives .dnnFormPopup the z-index). So, using the dialog directly, outside of the dnnModal function, has a broken display.

This commit moves the "z-index: 100000" style from .dnnFormPopup to .ui-dialog