DNN-4928 Fix error when editing JavaScript Library extension
When a JavaScript Library extension has a dependency on a library, and the
version of that dependency only has two digits (e.g. 1.7), then there is an
exception that is thrown when trying to format the version

This commit checks the version before formatting it, so that the exception
won't be thrown.