DNN-7144: removing 51Degrees dependency.
I created a new "ClientDependencyProvider" that replaced 51D's which uses the ASP.NET "HttpBrowserCapabilities" class.
To use it, there is a need to replace the 51D references in the distribution web.config file (which I didn't do as this should go into EVOQ codebase only and not Platform).

Note: some features that EVOQ Analytic is depending on from 51D (e.g., operating system) are not available through this provider and there is a need for a decision about them.