DNN-7742 fix error getting value from NameValueCollection
When getting a single value from a NameValueCollection, the collection is first
converted into an ILookup<string,string> and then converted into a
Dictionary<string,string>.  However, during the conversion to ILookup, any
single value is treated as a comma-delimited list, and so values with commas
produce multiple values for a single key in the ILookup.  When converting that
into a Dictionary, there is an error thrown, because there cannot be multiple
values per key in a Dictionary.

This commit adds an overload to ToLookup which takes a Boolean value,
indicating whether to treat single values as comma-delimited lists. It then
uses this overload, passing false, whenever GetValue or GetValueOrDefault is
called after ToLookup (as those two methods use ToDictionary under the hood).