Fix Bug7857 - WebAPI TabIsInPortal always returns true
fix to bug 7857,  before this would always return true regardless of the
portalId.

The method GetTab will work out the correct portalId to locate the tab
if the tabId doesn't existing in the portal. Therefore we need to test
to see if the returned TabInfo object has the portalId as the expected
value.