DNN-7941 - DAM module on Host File Management page shows error
The fix in the DNN Platform\DotNetNuke.Web\Api\Internal\DnnContextMessageHandler.cs TabIsInPortal method did in DNN-7857 does not handling properly host pages. Unfortunately there is no way in the DnnContextMessageHandler to know if a web method request is done in the context of an host tab. The portal id is based on the alias so this will always set a valid portal Id.
I have fixed extended the logic added in DNN-7857 returning true in the method if the tab is an Host tab. The private method has been renamed from *TabIsInPortal *to *TabIsInPortalOrHost*