DNN-8091 fix exception when settings not set
If the settings don't yet have a value, this throws an exception.  The intention was to throw only when the type or the attribute wasn't defined, but that wasn't what the implementation did.  I'm just going to leave the TODO comment for now to indicate that we should consider handling the possibility of other attributes somehow.