DNN-8867: improving folder permission caching.
Note: the solution still not optimal as there are no single calls to purge a specific older permissions from the Cache in the library and any change in the folders will purge all folders permissions through cache dependency. I tried to do this with minimal changes so as not to introduce new features and major changes to the system.

BTW, QA team must do some performance measurements on this change to see the effects on regular site operations.