Prevent unnecessary looping through cke settings
When there are a large number of portals, there is a severe performance
penalty when building the cke settings.  The first step to fixing this is
to limit the number of records being looped through to just the settings
matching the settings prefix when loading settings overrides for portals,
tabs, and modules instead of every record in the database table.  This has
a very large immediate performance improvement.