DNN-10082 Fix issue with JSON when sending message
The JSON sent to MessagingServiceController from ComposeMessage.js is invalid,
but fails in an acceptable way with Newtonsoft.Json v7. However, in v8 (through
the current version, v10), this causes a larger failure, which breaks the API
(resulting in a null reference exception at the start of the call).  This
change adjusts the JSON to always send an array (serialized as a string),
rather than sending an empty object when the array would be empty.  This also
requires the calling code to expect empty lists, rather than null lists.