DNN-10688 (#482)
* DNN-3468 (ex DNN-10688): Quickfix and unit test updates.
This bug is related to a design error in React/Redux, so a proper fix would consist in a refactoring of this part of the code. Current fix solve problem by hiding prompt commandline by CSS

* DNN-3468 (ex DNN-10688) Fix EOL

* DNN-10688 Fix EOL issue

* DNN-10688 Removed busy method. State change MUST be managed by Redux

* DNN-10688 Restore lost changes on Input.jsx

* DNN-10688 Introduced a patch to simulate desired behavior.
Note, comments added in Input.jsx. This feature should be achieved by a correct usage of Redux.

* DNN-10688 Fix conflicts

* DNN-10688: Fixed "dead x" character in pagination event