Content Localization: shared module un-published after changes (#152)
Clicking on an `Update localization`, application removes all modules for which `IsLocalized` property has been changed. Instead, system starts using reference to an original version of a module. Shared module is not getting re-published automatically, so it gets invisible in a view mode. As per customers, this is expected behavior as content of referenced module can be different.
Now users are confused when particular page settings are modified, but changes are applied to all other pages. This is because of shared module.
Customer don't want to have an effect globally and localization of a shared module property should be handled better, ideally, on a module level settings and not on a page settings. 
So, request is to add tooltip info whether module is shared or not and gray out (disable) the chain icon for a shared modules. 

fixes #151 & #152