Correction of .NET Framework Version Determination
Removed the lookups for prior versions of .NET Framework that are no longer possible for operation of DNN Platform after the upgrade to 4.7.2 as a platform requirement.

Renamed private static methods to properly indicate their dependence on a specific .NET Framework type.

Added support for the identification of .NET Framework version 4.8 display.  Additionally left a note for the reference list for future updates.