Standardization of Obsolete Flags
Updates across the entire repository to standardize the scheduled removal of API methods to adhere to the published Deprecation Policy as created by the TAG Team in mid-2018.  This aligns all items that will be removed as part of the 10.0 release to have a proper note.

It should be noted, that ALL items deprecated 8.x or prior have been updated to be removed as part of 10.0, this change is intentional and in aherenace to the published policy.  Additionally a blog post is being published today that outlines this change and the reasoning behind the change.