Limit rc action to run only on release branches
An issue was found where the action would run on any branch.
By mistake the filter was on just one of the steps
instead of the whole action.

This PR fixes that and applies the filter to the whole action.

This is a release management task and we are self-approving it for that reason