Restores icons in language editor
Closes #3754

There was a breaking change with I believe raw-loader or file-loader
or svg-loader or some such, we already had 2 other similar PRs for
this same issue in other areas.

After that dependency update, I think in 9.5.0,
we now need to add `.default` at the end of svg imports.

This fixes the same issue in the language resources editor.

I did a search to find if we missed this in other spots and did not find
any other instance.