Fixes a null reference exception with ISearchable (#3806)
Closes #3805

This is a followup on PR #3777

For context the deprecated class had the deprecated method required and the recommended one not required. PR #3777 deprecated the whole class for easier migration and reversed the required vs the non-required methods so developers can implement the new way without hacks or build warnings.

In that PR, a mistake was made on the deprecated method where we default to return null, but that caused a null reference exception on old modules. This PR replaces that return for an empty collection, which solves the same goal without an exception.