Added check for existing temp table
Just to make 100% sure it does not fail if there is a timeout we added a check for the existence of the temp table to delete it if present. This was discussed in the approvers meeting. This can be improved further as discussed but this PR is still an improvement so we will include it in 9.6.2