Resolves 9.6.2 logging issue.
Closes #3906

Apparently some files where deleted as part of #3843

This PR brings those files back and also fixes stylecop warnings. Some class names had to be changed both for clarity and to respect stylecop rule that the file name should match the class name. Because of that change, the dnn manifest was also changed so it all works with the new names.