Move Localization.SetThreadCultures to HttpModule
This is in support of turning on Async for Default.aspx. Per a
discussion with the ASP.NET team, setting the thread culture in OnInit
for a page does not persist in the context for the page (once Async is
turned on an there are multiple contexts to consider). Setting it from
an IHttpModule implementation, however, will persist throughout that
request.

This fixes the bug that plagued #2089, where, as an example, URLs would
be generated for the incorrect language once Async was on.