Support C# Async Constructs During Page Rendering (#3984)
* Use IPortalSettings in Localization APIs

This is in preparation to move the Localization.SetThreadCultures call
to an HttpModule.

* Move Localization.SetThreadCultures to HttpModule

This is in support of turning on Async for Default.aspx. Per a
discussion with the ASP.NET team, setting the thread culture in OnInit
for a page does not persist in the context for the page (once Async is
turned on an there are multiple contexts to consider). Setting it from
an IHttpModule implementation, however, will persist throughout that
request.

This fixes the bug that plagued #2089, where, as an example, URLs would
be generated for the incorrect language once Async was on.

* Add Async support to Default.aspx

This is a second try for 0cf1b13b2f076b6f7314142c1778ca97c267ba0b #2089.
This will allow for async calls from WebForms controls (preferably via
calling RegisterAsyncTask, see https://docs.microsoft.com/en-us/aspnet/web-forms/overview/performance-and-caching/using-asynchronous-methods-in-aspnet-45#registerasynctask-notes)