Fix UserRequestIPAddressController.cs IPv6 handling
GetUserRequestIPAddress has a bug since it assumes that all incoming requests are IPv4 addresses.  This causes bugs in all downstream IP address handling.  One specific example is the LOGIN_SUCCESS event log entry does not include the IPv6 address (IPv4 addresses are working).  My proposed change checks if the request is coming from a valid IPv4 address and returns it if it is IPv4.  If not, it checks if it is an IPv6 address and returns it or string.Empty if the IP address is not valid.