Prevents JWT upgrade to change existing config
With the recent change to make optional packages (.resources) upgrade if the package was previously installed, while upgrading the JWT authentication provider, it cause the existing web.config configuration to be reset to the default and that would happen on each upgrade.

This PR makes it ignore the node if it is already present from a previous installation.