Client side check for username minimum length
I added the configured value of UserNameMinLength to the application settings for the users module, and used that to verifiy the length client side. The message doesn't tell the user the exact minimum value, but the tooltip does now show that there is a check for the minimum configured length.