Update Compress Images GitHub Actions workflow
The Compress Images workflow isn't catching most of the images being
committed. This is, in part, because it doesn't run with permissions to
edit PRs on forks.  This updates the configuration to open a new PR when
a commit is merged, if there are images that can be compressed. It also
adds a scheduled run, in case something gets missed.
Based on https://github.com/calibreapp/image-actions#combined-workflow