Update RoleController.cs
This is to fix issue #2543.  It looked like the effective date and expiration date were switched; so I switched them back so the effective date was set to the current date; instead of the expiration date being set to today, thus having the user not added to the role when the rsvp code is entered.  This alone did not fix it though.  I'm unsure what the other if statement was for; but I just set to to set the value to Null.NullData as the ExpiryDate will always be < DateTime.Now since it is not defaulted to Null.  Hopefully someone else more familiar with this knows what this was supposed to do.  Was it meant for EffectiveDate too?