Refactor to support PersonaBar & EditBar theming
Since `PersonaBar` and `EditBar` are in two separate IFRAMEs, code needed to be refactored to support separate custom CSS files. Also, the purpose of this is to support custom `PersonaBar` and `EditBar` themes without having to modify core files. Custom theme files should be named appropriately (needs documentation) and placed in `/Portals/_default`.