Added lint task to all node projects
This adds a `lint` task into all projects with autofix enabled. This
will make it easier to auto-fix eslint warnings if we change the
configs or while updating eslint dependency in batch using lerna or yarn
workspaces. Unfortunatally (well fortunatelly) there was no autofixable
warnings to resolve after this change which was my main goal with this PR,
anyway, it's still good to have I believe for the future.