Added support for RelativeAppPath in Request token
Closes #5265

The .Net Request.ApplicationPath behaves this way:
- For a normal install it returns `/`
- For a virtual directory install it retursn `/dnn` (or whatever virtual directory name)

Notice here that in one case the string ends with a / and in the other case not.

This behaviour makes it diffucult to use it in a multi-purpose token made for relative urls as in:
```
<script src="[Request.ApplicationPath]/DesktopModules" />
```
Would render:
- `/dnn/DesktopModules` for a virtual directory (OK)
- `//DesktopModules` for a normal hosting scenario (NOT OK)

This PR adds `RelativeAppPath` which is a variation of `ApplicationPath` that will always end with `/`.

Usage as follows makes it usable in both scenarios and consumers can expect to always have the ending `/` provided:
```
<script src="[Request.RelativeAppPath]DesktopModules" />
```