Fixed an issue where first level folders were auto-expanded.
In #5351 a change was implemented to get the root folders according to the settings. Because of this a user is now always on some sort of root folder instead of the root folders not being part of any and having a virtual "root" ListItem. So before that change we were auto-expending that root item but now all direct children are open by default and it does not look right. This PR solves that by not expanding automatically the first level in the tree.