Fixed an issue that prevented sending smtp test emails
Fixed an issue that prevented sending smtp test emails
Due to the UI changes in #5152 the frontend now posts an obfuscated password unless the user is currently changing the password.
That scenario got handled correctly for UpdateSmtpSettings but the same logic needed to be implements in SendTestEmail for that button to work fine.