Fixed an issue that caused an infinite loop when saving pofiles
An automated stylecop fix caused a casing issue on a variable (IsValid vs isValid) that was a private field and a public property. The property calling a method and the method using the field (but in this case it was using the property by mistake and causing the endless loop).

I also fixed all other stylecop issues on the file and tested the build to confirm it fixes the issue.