Temp fix Resource Manager default height
Temporary fix for next patch release to give the Resource Manager a better default height.

In brief, the default height is currently hard-coded to 400px. This CSS-only fix instead bases the height on the browser's height using "100vh - 25rem" which basically means the resource manager's height will be (dynamically) about 400px less than the browser's height (to account for the theme's header and footer at least). We've also set a min-height of 30rem (appx 480px).

Read Issue #5361 for more details.
https://github.com/dnnsoftware/Dnn.Platform/issues/5361