Move DnnDeprecatedAttribute to Abstractions
In order for projects to not require a full reference to the source
generator project, the attribute cannot be defined in the project.
The first attempt to resolve this was to generate the attribute via the
source generator itself, but since the Library project exposes its
internals to other projects (e.g. DotNetNuke.HttpModules), this resulted
in ambiguous references. As a result, moving the type to a project that
all other projects can reference seemed to be the best compromise. Since
the attribute is conditional, it doesn't affect the public API of that
project.

Fixes #5815