Added sub-dependencies of AspNetWebApi to package
This may fix #6222 but I ran out of time to test, a clean install did work fine and that's all I had time to test for now.

I think there are a couple other dlls in the issue we could handle here or elsewhere and we need to test upgrade scenarios from 9.13.4,5,6 to be sure.