Updated Quick Settings Performance
- Fixes #6241

Given the reported issue in #6239, each time a module setting is updated the cache is cleared.  The current code, was always updating the first load to be false, even if the first load value was already set to false.

This change utilizes the retreived setting value to ensure that we are ONLY updateing the module to flip the quick settings flag to false if it is currently set to true.  This has a substantial impact to performance for modules using Quick Settings.