Added missing Styles.css file
Closes #6352

The persona bar loads up a css file by convention even though we don't need one for this module.

While I had my hands in there I notice another issues, the dnn-input controls have a floating label that assumes the field has the same background as the field. This was not the case, there was also an issue with dnn-elements with passing down colors to the child elements, so this both updates to the latest dnn-elements and adds a style to make the fields use the same background color as the surronding area.