Fixed a menu collapsion issue in Firefox
Firefox restores fields values upon going back and the mobile menu expand/collapse control is a checkbox restyled. This made it so upon reloading or going back in history in firefox, the menu would be collapsed but could show the icon as if it was expanded. This PR solves that.

Closes #6360