Refactored upgrade code after rollup of sql scripts
In #6038 a rollup script was put in place to revamp our clean installation logic, get rid of some upgrade logic and potentially speed up installs, etc.

In RC1 issue #6364 was reported. I spent quite a lot of time trying to figure it out. Had to do a `git bisect` and it identified the commit of the rollup script as being the first commit to cause that issue. After a lot of head scratching I realized that the code in upgrade.cs was the one creating the virtual host page that should hold the module that allows editing user profiles from the Persona Bar. This solved the issue for HOST users but not for admins. For admins, this lives in the portal templates, it was cleaned up from modules we don't want, but this one needed to be brought back as the Persona Bar needs it for admins.

As I was in upgrade.cs I though since we will document in our upgrade steps to upgrade to latest v9 before going to v10, we can get rid of all the other upgrade logic for all versions <10 in there.

Closes #6364